Added ability to refetch entrypoints during garden creation if they add one manually.
