+++ garden/package.json
@@ -47,7 +47,7 @@
     "sonner": "^1.5.0",
     "tailwind-merge": "^2.3.0",
     "tailwindcss-animate": "^1.0.7",
-    "vite": "^5.2.12",
+    "vite": "^5.3.3",
     "vite-tsconfig-paths": "^4.3.2",
     "web-vitals": "^4.0.1",
     "zod": "^3.23.8"
